Se agrego inicio dinamico al componente dateBar y se modifica el initialState para que incie en el mes actual

+++ src/Hooks/useBalance.js
@@ -42,7 +42,7 @@ function useBalance(dataMonth) {
         let total = 0;
 
         array.forEach((income) => {
-          total = total + parseInt(income.amount);
+          total = total + parseFloat(income.amount);
         });
 
         return total;
